TESTING/cobot-llm-control-v2/requirements.txt
groq==0.31.0
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0